requirements.txt
flask
fpdf
pillow
firebase-admin
PyMuPDF
